Detalles y rutas arreglados
