+++ src/views/partials/footer.ejs
@@ -2,7 +2,7 @@
     <div class="links">
         <a href="/">home</a>
         <a href="about">acerca de</a>
-        <a href="/menu">menu</a>
+        <a href="/producto/menu">menu</a>
         <a href="#">Contacto</a>
     </div>
