+++ src/views/partials/header.ejs
@@ -9,11 +9,11 @@
         <i class="fas fa-home"></i>
         <a href="#">Acerca De</a>
         <i class="fas fa-question-circle"></i>
-        <a href="/menu">Menu</a>
+        <a href="/producto/menu">Menu</a>
         <i class="fas fa-th-list"></i>
         <a href="#contact">Contacto</a>
         <i class="fas fa-address-book"></i>
-        <a href="/editar">Editar Productos</a>
+        <a href="/producto/editar">Editar Productos</a>
         <i class="fas fa-address-book"></i>
     </nav>
