+++ src/views/products/menu.ejs
@@ -21,16 +21,16 @@
             <div class="box-container">
                 <% for(let platillo of item){ %>
                 <div class="box">
-                    <img src="img/products/<%= platillo.image %>" alt="">
+                    <img src="/img/products/<%= platillo.image %>" alt="">
                     <h3><%= platillo.item%></h3>
-                    <h4><%= platillo.type%></h4>
-                    <div class="price"><%= platillo.price%>
+                    <!--<h4><%= platillo.type%></h4>-->
+                    <div class="price">$<%= platillo.price%>
                         <% if(platillo.discount != ''){ %>
                             <span><%= platillo.discount %></span>
                         <% }%>
                     </div> 
                     <a href="#" class="btn">agregar carrito</a>
-                    <a href="/producto/<%= platillo.id%>" class="btn">detalles</a>
+                    <a href="/producto/menu/<%= platillo.id%>" class="btn">detalles</a>
                 
                 </div> 
                 <% } %>
