+++ src/views/products/productosPorEditar.ejs
@@ -13,7 +13,7 @@
         <main>
             <section class="agregar-inventario">
 
-                <form action="/editar" method="POST" class="contenedor-formulario" enctype="multipart/form-data">
+                <form action="/producto/editar" method="POST" class="contenedor-formulario" enctype="multipart/form-data">
                     <h2>Formulario para nuevos productos</h2>
                     <label for="item">Nombre del producto</label>
                     <input class="item" type="text" name="item" id="item" placeholder="nuevo producto" required>
@@ -63,7 +63,7 @@
                 <div class="box-container">
                     <% for(let platillo of productos){ %>
                         <div class="box">
-                            <img src="img/products/<%= platillo.image %>" alt="">
+                            <img src="/img/products/<%= platillo.image %>" alt="">
                             <h3>Nombre: <%= platillo.item%>
                             </h3>
                             <h4>Tipo: <%= platillo.type%>
@@ -74,7 +74,7 @@
                             </h4>
                             <h4>Compras: <%= platillo.boughts%>
                             </h4>
-                            <div class="price">Precio:<%= platillo.price%>
+                            <div class="price">Precio: $<%= platillo.price%>
                                     <% if(platillo.discount !='' ){ %>
                                         <span>
                                             <%= platillo.discount %>
@@ -82,10 +82,10 @@
                                         <% }%>
                             </div>
                             <button class="btn" type="button"
-                                onClick="location.href='./editarProducto/<%=platillo.id%>'">
+                                onClick="location.href='/producto/editarProducto/<%=platillo.id%>'">
                                 Editar Producto
                             </button>
-                            <form action="/delete/<%=platillo.id%>?_method=DELETE" method="POST">
+                            <form action="/producto/delete/<%=platillo.id%>?_method=DELETE" method="POST">
                                 <button type="submit" class="btn">ELIMINAR</button>
                             </form>
